requirements.txt
alembic==1.7.4
appdirs==1.4.4
attrs==21.2.0
bcrypt==3.2.0
black==21.7b0
cffi==1.14.6
click==8.0.1
Flask==2.0.1
flask-expects-json==1.6.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
importlib-metadata==4.8.1
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.1
jsonify==0.5
jsonpickle==2.0.0
jsonschema==4.1.2
Mako==1.1.5
MarkupSafe==2.0.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.26
pathspec==0.9.0
protobuf==3.17.3
pycparser==2.20
pyrsistent==0.18.0
python-dotenv==0.19.0
regex==2021.7.6
six==1.16.0
SQLAlchemy==1.4.22
tomli==1.1.0
Werkzeug==2.0.1
zipp==3.6.0
